# src/main/resources/openapiserver.yaml
swagger: "2.0"
info:
  version: 1.0.0
  title: workflow API
  description: APIs to interact with account objects
  contact: { }
schemes:
  - "http"
  - "https"
host: "localhost"
basePath: /
...
paths:
  /startProcess:
    get:
      produces:
        - "application/json"
      tags:
        - Workflow
      operationId: startProcessUsingGET
      responses:
        200:
          description: User event saved successfully.
          schema:
            $ref: "#/definitions/Process"
        400:
          description: User event couldn't have been saved.
        500:
          description: An unexpected error occured.
          schema:
            $ref: "#/definitions/Error"
...
definitions:
  Process:
    properties:
      processId:
        type: string
        format: UUID
      processName:
        type: string
      processVersion:
        type: string
      processStatus:
        type: string
      currentStep:
        type: string
      parentProcessId:
        type: string
      creationTime:
        type: string
        format: date-time
      currentStepEntryTime:
        type: string
        format: date-time
      closingTime:
        type: string
        format: date-time
      updateTime:
        type: string
        format: date-time
...
  Error:
    properties:
      code:
        type: string
      message:
        type: string